back-end/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.3
